Creates the node project and adds sequelize models
The project is coded with typescript in the src folders and then
compiled to the dist folder.

You must compile the typescript files before running the project. This
can be easily done by running "npm start".

The communication with the database is done with Sequelize. The models
and associations have been done but the typescript classes will maybe
be modified in the future (by specifying the foreign keys and/or
objects in fields).