Fix request problems between the server and client
Added the body parser middleware to receive the data (json or url
encoded) from the client. Without this middleware the body was always
empty.

Added an app middleware who add the CORS header to each respond of the
server. Without that the client can't receive the data from the server.

Added a method to return and log the errors encountered. Not the best
pratices (the messages are written in french to display them directly to
the client) but it should help a bit more in case of error.